Add support for using the quickfix window
Instead of opening the location list for the current window it should be
possible to use the quickfix window instead. This adds a configuration
value that is initialized to reflect the old behaviour.

By adding `let g:cscope_use_quickfix = 1` the quickfix window is used
instead of the location list for the current window.